Migrating victory-errorbar to TypeScript
