+++ packages/victory-core/src/victory-util/add-events.tsx
@@ -62,6 +62,7 @@ export interface EventsMixinClass<TProps> {
   ): React.ReactElement;
   cacheValues<TThis>(this: TThis, obj: Partial<TThis>): void;
   getEventState: typeof Events.getEventState;
+  renderData(props: TProps);
   renderContinuousData(props: TProps);
   animateComponent(props: TProps, defaultAnimationWhitelist);
   getComponentProps(
