--- packages/victory-errorbar/src/index.js
@@ -1,2 +1,0 @@
-export { default as VictoryErrorBar } from "./victory-errorbar";
-export { default as ErrorBar } from "./error-bar";
