+++ packages/victory-errorbar/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./victory-errorbar";
+export * from "./error-bar";
